Support manual vs generated stubs again

+++ maya_stubgen/generate_stubs/__init__.py
@@ -6,23 +6,25 @@ import traceback
 from importlib import import_module
 from pathlib import Path
 from pkgutil import ModuleInfo, walk_packages
+from textwrap import dedent
 from typing import *
 
 import click
 
-from .common import STUB_HEADER, Class, Function, FunctionType, Variable
+from .common import STUB_HEADER, Class, Function, Variable
 from .generate_cmds_stubs import generate_cmds_stubs
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 
-def get_stub_path(module: ModuleInfo):
+def get_stub_path(module: ModuleInfo, generated=False):
     stub_path = Path(module.name.replace("maya", "maya-stubs", 1).replace(".", "/"))
+    suffix = "_generated" if generated else ""
     if module.ispkg:
-        stub_path = stub_path / "__init__.pyi"
+        stub_path = stub_path / f"__init__{suffix}.pyi"
     else:
-        stub_path = stub_path.with_suffix(".pyi")
+        stub_path = stub_path.with_name(f"{stub_path.name}{suffix}.pyi")
     return stub_path
 
 
@@ -52,15 +54,23 @@ def generate_generic_stub(module: ModuleInfo) -> str:
 def generate_stubs_for_module(module: ModuleInfo):
     logger.info("Generating stubs for: %s", module.name)
 
-    stub_path = get_stub_path(module)
     if module.name == "maya.cmds":
         content = generate_cmds_stubs()
     else:
         content = generate_generic_stub(module)
 
+    stub_path = get_stub_path(module)
+    stub_path_generated = get_stub_path(module, generated=True)
+
+    # stub_path and stub_path_generated have the same parent
     stub_path.parent.mkdir(parents=True, exist_ok=True)
 
-    with stub_path.open("w", encoding="utf8") as f:
+    # Create the manually edited stub file and import the generated one in.
+    if not stub_path.exists():
+        with stub_path.open("w", encoding="utf8") as f:
+            f.write("# fmt: off\n" f"from .{stub_path_generated.stem} import *\n")
+
+    with stub_path_generated.open("w", encoding="utf8") as f:
         f.write(content)
